+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [6.0.9] - 2026-04-19
+
 ### Changed
 
 - Changed init from using on (svelte/events) to addEventListener